+++ maps/map7
@@ -1,30 +1,26 @@
 500
 ##start
-start 50 500
+start 50 400
+1 500 100
 4 200 100
-5 850 100
+5 800 100
 6 200 200
-7 850 200
-8 150 300
-9 850 300
+7 800 200
+8 200 300
+9 800 300
 #visual
-10 150 400
-11 850 400
-12 150 500
-13 850 500
-14 150 600
-15 850 600
-16 150 700
-17 850 700
-18 200 800
-19 850 800
+10 200 400
+11 800 400
+12 200 500
+13 800 500
+14 200 600
+15 800 600
+16 200 700
+17 800 700
 20 200 900
-21 850 900
-22 100 850
-23 150 950
-24 900 950
+21 800 900
 ##end
-end 1000 500
+end 950 400
 start-4
 start-6
 start-8
@@ -32,20 +28,14 @@ start-10
 start-12
 start-14
 start-16
-start-18
-start-20
-start-22
-4-5
+4-1
+1-5
 6-7
 8-9
 10-11
 12-13
 14-15
 16-17
-18-19
-20-21
-22-23
-23-24
 end-5
 end-7
 end-9
@@ -53,6 +43,3 @@ end-11
 end-13
 end-15
 end-17
-end-19
-end-21
-end-24
