+++ maps/map8
@@ -1,0 +1,36 @@
+20
+##start
+start 100 400
+1 100 60
+2 400 60
+3 700 60
+5 250 200
+6 550 200
+4 250 400
+7 550 400
+8 700 400
+##end
+end 900 400
+10 400 700
+#visual
+12 550 600
+11 550 700
+13 900 700
+start-1
+1-2
+5-6
+7-8
+12-end
+end-13
+end-8
+7-12
+10-4
+11-10
+3-8
+10-12
+6-7
+13-11
+7-4
+2-3
+start-5
+4-start
