+++ output
@@ -1,6 +1,11 @@
-L1-5 L2-4 L3-8 
-L1-7 L2-7 L3-9 L4-5 L5-4 L6-8 
-L3-10 L4-7 L5-7 L6-9 L7-5 L8-4 L9-8 
-L5-7 L6-10 L7-7 L8-7 L9-9 L10-5 
-L8-7 L9-10 L10-7 
-L10-7 
+L1-4 L2-1 L3-5 
+L1-10 L2-2 L3-6 L4-4 L5-1 L6-5 
+L1-11 L2-3 L3-7 L4-10 L5-2 L6-6 L7-4 L8-1 L9-5 
+L1-13 L2-8 L3-12 L4-11 L5-3 L6-7 L7-10 L8-2 L9-6 L10-4 L11-1 L12-5 
+L1-end L2-end L3-end L4-13 L5-8 L6-12 L7-11 L8-3 L9-7 L10-10 L11-2 L12-6 L13-4 L14-1 L15-5 
+L4-end L5-end L6-end L7-13 L8-8 L9-12 L10-11 L11-3 L12-7 L13-10 L14-2 L15-6 L16-4 L17-1 L18-5 
+L7-end L8-end L9-end L10-13 L11-8 L12-12 L13-11 L14-3 L15-7 L16-10 L17-2 L18-6 L19-4 L20-1 
+L10-end L11-end L12-end L13-13 L14-8 L15-12 L16-11 L17-3 L18-7 L19-10 L20-2 
+L13-end L14-end L15-end L16-13 L17-8 L18-12 L19-11 L20-3 
+L16-end L17-end L18-end L19-13 L20-8 
+L19-end L20-end
